Fix API compatibility, client data coherence, and add Docker data viewer
- Fix HEAD request 403: change all routes to Method: ANY + add {proxy+} catch-all
- Add org-level device stats endpoint (GET /api/v1/orgs/{org_id}/stats/devices)
- Fix client MAC/hostname coherence: unified CLIENT_PROFILES ensure hostname,
  OUI, manufacturer, OS, family, model are always consistent
- Separate wireless/wired weight pools (no phones wired, no desktops wireless)
- Fix band field "2.4" to "24" per Mist OpenAPI spec
- Add missing wireless client fields (rx_rate, tx_rate, rx_retries, tx_retries)
- Add wired client hostname and auth_method fields
- Fix client IPs to use site_index for subnet alignment
- Add Docker data viewer (iracic82/mist-data-viewer) with coherence validation UI
- Update README with Docker instructions and screenshot

+++ seed_data/generators/network_generator.py
@@ -62,8 +62,8 @@ class NetworkGenerator:
             },
             "vpn_access": {},
             "tenants": {},
-            "created_time": created,
-            "modified_time": now - random.randint(0, 86400 * 7),
+            "created_time": float(created),
+            "modified_time": float(now - random.randint(0, 86400 * 7)),
         }
 
     def generate_org_networks(self, org_id: str, seed: int = 42) -> list[dict]:
@@ -122,8 +122,8 @@ class NetworkGenerator:
             },
             "vpn_access": {},
             "tenants": {},
-            "created_time": created,
-            "modified_time": now - random.randint(0, 86400 * 7),
+            "created_time": float(created),
+            "modified_time": float(now - random.randint(0, 86400 * 7)),
         }
 
     def generate_derived_networks_for_site(
